Moved cardinal_of.hpp
