+++ include/boost/simd/detail/pack_info.hpp
@@ -12,7 +12,7 @@
 #ifndef BOOST_SIMD_DETAIL_PACK_INFO_HPP_INCLUDED
 #define BOOST_SIMD_DETAIL_PACK_INFO_HPP_INCLUDED
 
-#include <boost/simd/cardinal_of.hpp>
+#include <boost/simd/meta/cardinal_of.hpp>
 #include <boost/dispatch/meta/downgrade.hpp>
 #include <boost/dispatch/meta/upgrade.hpp>
