+++ include/boost/simd/meta/cardinal_of.hpp
